+++ client/src/pages/AccountsApproval/AccountsApproval.scss
@@ -1,12 +1,5 @@
 @import '../../scssStyles/mixins';
 
-.gbjgbjfbgjfbg {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  // background: white;
-}
-
 .accounts-approval-page-container {
   position: absolute;
   display: flex;
@@ -103,7 +96,7 @@
   justify-content: center;
   align-items: center;
   height: fit-content;
-  background: white;
+  background: var(--white);
   border-radius: 10px;
 }
 .all-accounts-buttons {
